requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.1
requests==2.26.0